+++ punchout_product_brand/__manifest__.py
@@ -1,0 +1,17 @@
+# Copyright 2025 Bosd
+# License AGPL-3.0 or later (https://www.gnu.org/licenses/agpl).
+
+{
+    "name": "Punchout Product Brand",
+    "version": "19.0.1.0.0",
+    "license": "AGPL-3",
+    "summary": "Map a punchout cart field onto the product brand",
+    "author": "OBS Solutions Netherlands, Odoo Community Association (OCA)",
+    "maintainers": ["bosd"],
+    "website": "https://github.com/OCA/edi",
+    "depends": [
+        "punchout_purchase",
+        "product_brand",
+    ],
+    "auto_install": True,
+}