+++ punchout_product_brand/models/__init__.py
@@ -1,0 +1,5 @@
+# Copyright 2025 Bosd
+# License AGPL-3.0 or later (https://www.gnu.org/licenses/agpl).
+
+from . import punchout_field_mapping
+from . import punchout_backend