+++ punchout_purchase/models/__init__.py
@@ -3,6 +3,7 @@
 # Copyright 2025 Bosd
 # License AGPL-3.0 or later (https://www.gnu.org/licenses/agpl).
 
+from . import punchout_field_mapping
 from . import punchout_backend
 from . import punchout_session
 from . import purchase_order